+++ src/proto/middleware.rs
@@ -1,5 +1,5 @@
 #![macro_use]
-use ArcCore::{Request, Response};
+use core::{Request, Response};
 use std::sync::Arc;
 use futures::future::{Future, IntoFuture};
